Moved collision detection and gravity to game.cpp

+++ aldebaran/source/headers/vector_2.h
@@ -4,7 +4,7 @@ struct Vector2 {
 	float x;
 	float y;
 	Vector2() : x(0), y(0) {}
-	Vector2(int x, int y) : x(x), y(y) {}
+	Vector2(float x, float y) : x(x), y(y) {}
 	Vector2 zero() {
 		return Vector2(0, 0);
 	}
